// src/api/models/createMRQRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */
import type { MRQData } from './mRQData';

/**
 * Schema for creating a new MRQ question.
 */
export interface CreateMRQRequest {
  template_id?: string | null;
  /**
   * @minLength 1
   * @maxLength 5000
   */
  question_text: string;
  question_type?: 'mrq';
  data: MRQData;
}