// src/api/models/createShortAnswerRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */
import type { ShortAnswerData } from './shortAnswerData';

/**
 * Schema for creating a new short answer question.
 */
export interface CreateShortAnswerRequest {
  template_id?: string | null;
  /**
   * @minLength 1
   * @maxLength 5000
   */
  question_text: string;
  question_type?: 'short_answer';
  data: ShortAnswerData;
}